Allow partly parallel Maven execution
This partly fixes Maven execution in parallel forks, because it postpones
evaluating of the rootFolder (`user.dir` value) until it is really needed
during execution. Without the fix the path could be evaluated in the main
process, meaning the `user.dir` is not set correctly to the tested module.

It still does not allow reusing of the forks (reuseForks parameter in
maven-surefire-plugin and maven-failsafe-plugin), but at least allows
parallel execution (forkCount greater than 1).

When reusing the forks, the SnapshotSystemJUnit5.finishedAllTests()
complains that it was called more than once, so the reuseForks has to
be disabled (set to false).

This partly fixes #551.

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>